+++ src/java_core_final_project/model/ParsedTransferData.java
@@ -3,5 +3,5 @@ package java_core_final_project.model;
 public record ParsedTransferData(
         String fromAccount,
         String toAccount,
-        int amount
+        double amount
 ) {}
